requirements.txt
# ===== 核心Web框架 =====
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-Caching==2.1.0
Flask-Mail==0.9.1
Werkzeug==3.0.1
SQLAlchemy==2.0.23

# ===== 数据库驱动 =====
PyMySQL==1.1.0
mysqlclient==2.2.0

# ===== 异步和并发支持 =====
aiohttp==3.9.1
aiofiles==23.2.0
asyncio-mqtt==0.16.1
nest-asyncio==1.5.8

# ===== 服务器和部署 =====
gunicorn==21.2.0
uvicorn[standard]==0.24.0
gevent==23.9.1

# ===== 翻译和AI相关 =====
dashscope==1.17.0
openai==1.6.1
httpx==0.25.2

# ===== PPT和文档处理 =====
python-pptx==0.6.23
openpyxl==3.1.2

# ===== PDF处理 =====
pymupdf==1.24.9
pdfplumber==0.11.2
pdf2docx==0.5.8
python-docx==1.1.0
pypandoc==1.13

# ===== OCR和图像处理 =====
easyocr==1.7.2
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.4

# ===== 网络爬虫和数据处理 =====
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.0
lxml==4.9.3

# ===== 系统监控和性能 =====
psutil==5.9.6
GPUtil==1.4.0
prometheus-client==0.19.0

# ===== 工具库 =====
python-dotenv==1.0.0
pytz==2023.3
tqdm==4.66.1
click==8.1.7
colorama==0.4.6

# ===== 日志和调试 =====
loguru==0.7.2
concurrent-log-handler==0.9.25

# ===== 数据验证和序列化 =====
marshmallow==3.20.1
pydantic==2.5.0

# ===== 缓存 =====
redis==5.0.1

# ===== 安全 =====
cryptography==41.0.7
bcrypt==4.1.2

# ===== 开发和测试工具 =====
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# ===== FastAPI (可选，用于API服务) =====
fastapi==0.104.1
python-multipart==0.0.6

# ===== 其他依赖 =====
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
blinker==1.7.0